fix(session): address code review feedback
